+++ tests/php/TestCase.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace OmniTend\LaravelDashboard\Tests;
+
+use OmniTend\LaravelDashboard\LaravelDashboardServiceProvider;
+use Orchestra\Testbench\TestCase as Orchestra;
+
+abstract class TestCase extends Orchestra
+{
+    protected function getPackageProviders($app): array
+    {
+        return [
+            LaravelDashboardServiceProvider::class,
+        ];
+    }
+
+    protected function defineEnvironment($app): void
+    {
+        $app['config']->set('database.default', 'testing');
+        $app['config']->set('database.connections.testing', [
+            'driver' => 'sqlite',
+            'database' => ':memory:',
+            'prefix' => '',
+        ]);
+    }
+}